Move back TestStorageMonitoring
